requirements.txt
annotated-types==0.7.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
iniconfig==2.1.0
numpy==2.0.2
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
